most tests passing. need to try reducing concurrency as it looks like tests are racing each other and failing intermittently

+++ actions/delete-file/index.js
@@ -33,8 +33,13 @@ async function main(params) {
 
     var response = {};
     try {
-      await files.delete(params.target);
-      response["statusCode"] = 200;
+      var props = files.getProperties(params.target)
+      if (!props) {
+        return errorResponse(404, "NotFound", logger);
+      } else {
+        await files.delete(params.target);
+        response["statusCode"] = 200;
+      }
     } catch (error) {
       logger.info("caught error:")
       return errorResponse(400, error, logger);
